Update The_Great_Weather_Forcast_App.py
This is my weather forecast. With this Forcast, you can check what is the weather in the city you wrote in the text box, the description of the city, the temperature, the minimum temperature, and the wind speed there. If you type a random thing that is not a city or the wrong spelling of the city, it will say, "Please provide a valid city name". I will update this code to add more features. Meanwhile I update, please code and enjoy!!